Move Hudi class name constants to HiveClassNames
